Moved deformation variables classes to own module.
